feat: improve mobile layout & fix footer issue
